docker/airflow/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-docker
astronomer-cosmos==1.7.1
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-openlineage==1.10.0
openmetadata-ingestion
# apache-airflow[http]
dbt-postgres
